simplify, move datasource to separate class
