+++ projects/ui-components/lf-selection-list/lf-selection-list-public-api.ts
@@ -1,3 +1,4 @@
 export * from './lf-selection-list.component';
+export * from './lf-selection-list-types';
 
 export * from './lf-selection-list.module';
