+++ projects/ui-components/lf-selection-list/lf-selection-list-types.ts
@@ -1,0 +1,18 @@
+import { ILfSelectable } from '@laserfiche/lf-ui-components/shared';
+
+export interface ColumnOrderBy {
+  columnId: string;
+  isDesc: boolean;
+}
+
+export interface ColumnDef {
+  id: string;
+  displayName: string;
+  defaultWidth: string;
+}
+
+/** @internal */
+export interface SelectedItemEvent {
+  selected: ILfSelectable;
+  selectedItems: ILfSelectable[] | undefined;
+}
