+++ projects/ui-components/lf-selection-list/lf-selection-list.component.html
@@ -1,6 +1,7 @@
 <cdk-virtual-scroll-viewport
   #viewport
   id="lf-list-viewport"
+  tabindex="0"
   [itemSize]="itemSize"
   (scrolledIndexChange)="onScroll($event)"
   (keydown)="onViewportKeyDown($event)"
@@ -13,24 +14,19 @@
     class="lf-table-selection-list"
     matSort
     (matSortChange)="sortData($event)"
-    [ngClass]="{
-      'table-width-auto': !(columns?.length > 0)
-    }"
     multiTemplateDataRows="true"
   >
     <ng-container [matColumnDef]="column.id" *ngFor="let column of allColumnDefs; let i = index">
       <ng-container [ngSwitch]="column.id">
-        <!-- the same view can be shown in more than one case -->
         <ng-container *ngSwitchCase="'name'">
           <th mat-header-cell *matHeaderCellDef class="lf-ellipsis-overflow">
-            <!-- <span class="resize-handle resize-handle-left" (mousedown)="onResizeColumn($event, multipleSelection ? 0 : 0)"></span> -->
             <div class="lf-resize-wrapper-div">
-              <div class="column-text-div" mat-sort-header>Name</div>
-              <span class="resize-handle resize-handle-right" (mousedown)="onResizeColumn($event, i)"></span>
+              <div class="column-text-div" mat-sort-header>{{column.displayName}}</div>
+              <span class="resize-handle" (mousedown)="onResizeColumn($event, i)"></span>
             </div>
           </th>
           <td mat-cell *matCellDef="let item">
-            <div class="icon-name-container">
+            <div class="lf-name-col-container">
               <ng-container *ngIf="listItemRef; else defaultItem"
                 ><ng-container *ngIf="item">
                   <ng-container *ngTemplateOutlet="listItemRef; context: item"></ng-container>
@@ -66,12 +62,11 @@
           <th mat-header-cell *matHeaderCellDef class="lf-ellipsis-overflow">
             <div class="lf-resize-wrapper-div">
               <div class="column-text-div" mat-sort-header>
-                <!-- <span class="resize-handle resize-handle-left" (mousedown)="onResizeColumn($event, i - 1)"></span> -->
                 {{ column.displayName }}
               </div>
               <span
                 *ngIf="i > 0"
-                class="resize-handle resize-handle-right"
+                class="resize-handle"
                 (mousedown)="onResizeColumn($event, i)"
               ></span>
             </div>
